data_ingestion_crawler/requirements.txt
fastapi
uvicorn
python-dotenv
azure-storage-blob
azure-ai-documentintelligence
azure-cosmos
openai
langchain
tiktoken
